Start conversion of tracer settings
