+++ src/main/java/net/countercraft/movecraft/combat/features/tracers/commands/TracerSettingCommand.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.combat.commands;
+package net.countercraft.movecraft.combat.features.tracers.commands;
 
 import net.countercraft.movecraft.combat.MovecraftCombat;
 import net.countercraft.movecraft.combat.localisation.I18nSupport;
@@ -12,10 +12,10 @@ import java.util.List;
 
 import static net.countercraft.movecraft.util.ChatUtils.MOVECRAFT_COMMAND_PREFIX;
 
-public class TracerModeCommand implements TabExecutor {
+public class TracerSettingCommand implements TabExecutor {
     @Override
     public boolean onCommand(CommandSender commandSender, Command command, String s, String[] args) {
-        if(!command.getName().equalsIgnoreCase("tracermode"))
+        if(!command.getName().equalsIgnoreCase("tracersetting"))
             return false;
 
         if(!(commandSender instanceof Player)) {
@@ -25,22 +25,22 @@ public class TracerModeCommand implements TabExecutor {
         Player player = (Player) commandSender;
 
         if(args.length == 0) {
-            commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Command - Current Mode") + ": " + MovecraftCombat.getInstance().getPlayerManager().getMode(player));
+            commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Command - Current Setting") + ": " + MovecraftCombat.getInstance().getPlayerManager().getSetting(player));
             return true;
         }
         if(args.length != 1) {
-            commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Command - Specify Mode"));
+            commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Command - Specify Setting"));
             return true;
         }
 
-        String mode = args[0].toUpperCase();
-        if(!mode.equals("BLOCKS") && !mode.equals("PARTICLES")) {
-            commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Command - Specify Valid Mode"));
+        String setting = args[0].toUpperCase();
+        if(!setting.equals("OFF") && !setting.equals("LOW") && !setting.equals("MEDIUM") && !setting.equals("HIGH")) {
+            commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Command - Specify Valid Setting"));
             return true;
         }
 
-        MovecraftCombat.getInstance().getPlayerManager().setMode(player, mode);
-        commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Command - Tracer Set") + ": " + mode);
+        MovecraftCombat.getInstance().getPlayerManager().setSetting(player, setting);
+        commandSender.sendMessage(MOVECRAFT_COMMAND_PREFIX + I18nSupport.getInternationalisedString("Command - Tracer Set") + ": " + setting);
         return true;
     }
 
@@ -48,8 +48,10 @@ public class TracerModeCommand implements TabExecutor {
     public List<String> onTabComplete(CommandSender sender, Command command, String s, String[] strings) {
         final List<String> tabCompletions = new ArrayList<>();
         if (strings.length <= 1) {
-            tabCompletions.add("BLOCKS");
-            tabCompletions.add("PARTICLES");
+            tabCompletions.add("OFF");
+            tabCompletions.add("MEDIUM");
+            tabCompletions.add("HIGH");
+            tabCompletions.add("LOW");
         }
         if (strings.length == 0) {
             return tabCompletions;
